+++ src/ssh/azext_ssh/tests/latest/test_resource_type_utils.py
@@ -12,8 +12,9 @@ from azext_ssh import ssh_info
 
 from azure.cli.core import azclierror
 
+
 class SshResourceTypeUtilsCommandTest(unittest.TestCase):
-    
+
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_ip(self, mock_list_types):
         cmd = mock.Mock()
@@ -21,22 +22,21 @@ class SshResourceTypeUtilsCommandTest(unittest.TestCase):
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.Compute/virtualMachines")
         mock_list_types.assert_not_called()
 
-    ################################ Test Resource Type Provided ##############################
-    
+    # Test Resource Type Provided
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_arc(self, mock_list_types):
         cmd = mock.Mock()
         mock_list_types.return_value = {'microsoft.hybridcompute/machines', 'microsoft.compute/virtualmachines'}
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.HybridCompute/machines", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.HybridCompute/machines")
-    
+
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_compute(self, mock_list_types):
         cmd = mock.Mock()
         mock_list_types.return_value = {'microsoft.hybridcompute/machines', 'microsoft.compute/virtualmachines', 'microsoft.connectedvmwarevsphere/virtualmachines'}
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.Compute/virtualMachines", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.Compute/virtualMachines")
-    
+
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_vmware(self, mock_list_types):
         cmd = mock.Mock()
@@ -44,22 +44,21 @@ class SshResourceTypeUtilsCommandTest(unittest.TestCase):
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.connectedvmwarevsphere/virtualMachines", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.ConnectedVMwarevSphere/virtualMachines")
 
-
-    ############################ Test Legacy Resource Type (Resource Provider) ##########################
+    # Test Legacy Resource Type (Resource Provider)
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_arc_legacy(self, mock_list_types):
         cmd = mock.Mock()
         mock_list_types.return_value = {'microsoft.hybridcompute/machines', 'microsoft.compute/virtualmachines'}
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.HybridCompute", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.HybridCompute/machines")
-    
+
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_compute_legacy(self, mock_list_types):
         cmd = mock.Mock()
         mock_list_types.return_value = {'microsoft.hybridcompute/machines', 'microsoft.compute/virtualmachines', 'microsoft.connectedvmwarevsphere/virtualmachines'}
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.Compute", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.Compute/virtualMachines")
-    
+
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_resourcetype_vmware_legacy(self, mock_list_types):
         cmd = mock.Mock()
@@ -67,7 +66,7 @@ class SshResourceTypeUtilsCommandTest(unittest.TestCase):
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, "Microsoft.connectedvmwarevsphere", None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.ConnectedVMwarevSphere/virtualMachines")
 
-    ############################# Test No Resource Type Provided ###################################
+    # Test No Resource Type Provided
     @mock.patch('azext_ssh.resource_type_utils._list_types_of_resources_with_provided_name')
     def test_decide_resource_type_more_than_one(self, mock_list_types):
         cmd = mock.Mock()
@@ -105,6 +104,5 @@ class SshResourceTypeUtilsCommandTest(unittest.TestCase):
         op_info = ssh_info.SSHSession("rg", "vm", None, None, None, False, None, None, None, None, [], False, None, None, None, False, False)
         self.assertEqual(resource_type_utils.decide_resource_type(cmd, op_info), "Microsoft.ConnectedVMwarevSphere/virtualMachines")
 
-    
     if __name__ == '__main__':
         unittest.main()